chore(#3779): connection query cleanup | extract base query builder
### Summary

Extracts the two large base-query-building methods (`BuildBaseQueryFromOthers` / `BuildBaseQueryToOthers`) and their supporting code out of `ConnectionQuery` into dedicated files, improving separation of concerns and readability.

### Changes

**New files:**

- **`ConnectionBaseQueryBuilder.cs`** — New `internal` class containing:
  - `FromOthers()` — builds the base `IQueryable` for "connections from others" (direct assignments, key-role expansion, role-map joins, delegations, child/hierarchy nesting, innehaver connections)
  - `ToOthers()` — builds the base `IQueryable` for "connections to others" (direct + main-unit assignments, role-map joins, delegations, key-role expansion)
  - `BuildFilterSets()` — shared helper that constructs the filter `HashSet`s (via, viaRole, role, roleExclude) used by both directions
  - `PostgresPlanHintTakeLimit` constant and `ReviRegnRoleSet` — moved from `ConnectionQuery`

- **`ConnectionQueryExtensions.cs`** — Extension methods (`ToIdContains`, `FromIdContains`, `ViaIdContains`, `RoleIdContains`, `RoleIdExcludes`, `ViaRoleIdContains`) extracted from the bottom of `ConnectionQuery` into their own file

- **`ConnectionQueryDirection.cs`** — `ConnectionQueryDirection` enum (`FromOthers`, `ToOthers`) extracted into its own file

**Modified files:**

- **`ConnectionQuery.cs`** — Removed the extracted methods, constants, extensions, and enum. Now instantiates `ConnectionBaseQueryBuilder` and delegates base-query construction to it. All public API signatures remain unchanged.

### No behavioral changes

This is a pure structural refactor. No SQL/query logic was modified. Existing integration tests pass without changes.

Closes #3779 — as a part of #3775.
